Remove SearchStrategy.CURRENT
This was set on the ChannelBindingProperties bean. However, since the ModuleProperties
(which is of type ChannelBindingProperties) might have been initialized by the initializer,
the search strategy for the conditional bean needs to look at the hierarchy.

Added a TODO comment indicating this can be set to CURRENT once the initializer is removed.