Re-instate search=CURRENT for config properties
In order to run a module as a child context (namespaceing the channels)
you need to have a separate ChannelBindingProperties per context,
so the condition has to have search=CURRENT. For an XD module this
was a problem because the ModuleProperties (a subclass) is needed and that
has to be added in bootstrap (i.e. in the parent context). Fixed by
registering the ModuleProperties as a singleton in the child context.