Fix BindingServiceConfiguration for SI-5.2
Starting with version 5.2 Spring Integration provides a global
`DestinationResolver` bean for some runtime optimization using a single
object.
With that the `binderAwareRouterBeanPostProcessor` bean can't resolve
a single `DestinationResolver` bean any more since there is now two
instances of that.

* Fix `BindingServiceConfiguration.binderAwareRouterBeanPostProcessor()`
to use `@Qualifier("binderAwareChannelResolver")` for an explicit bean
resolution for injection.