GH-1893 Fix SpEL partition support
Consolidated code around PartitionHandler and interceptor

Consolidated ContentTypeConfiguration and the returning CompositeMessageConverter

Added support (reluctantly) to allow SpEL expressions for determining partition for function programming model. In other words something like  would work

That said, it is highly recommended to never use payload as a source for any kind of expressions (input or output) as payload constitutes privileged information that should only be accessible to he producer and consumer.

Resolves #1893
Resolves #1877