Reactive functions and native encoding
When native encoding is enabled with a reactive function,
Spring Cloud Stream is still trying to convert the message
causing ClassCastException and other potential issues downstream.
Fixing this issue by signaling the framework to not perform any
message conversion on the outbound when useNativeEncoding is
enabled with a reactive function.

Resovles https://github.com/spring-cloud/spring-cloud-stream/issues/2037