GH-2129 Ensure exception is propagated from ApplicationJsonMessageMarshallingConverter
This will ensure that any exception that is thrown during the conversion by ApplicationJsonMessageMarshallingConverter are propagated. Basically the idea is that since MessageConverter has accepted the conversion via canConvert() method it is no longer necessary to expect that may be some other converter in the stack may succeed. And on top of that user can still define custom message converter

Resolves #2129
Resolves #2191