Add `BindingService.get*Binding(s)(String)` API
There are some use-cases when we need to know upfront
if specific binding is really bound.
The goal is to prevent some security vulnerabilities
when too many unexpected dynamic bindings, e.g.
via router based on the `StreamBridge`.
The `BindingService` provides for us `getProducerBindingNames()`
 and `getConsumerBindingNames()` which are not optimal
 for the task to check for binding presence

* Introduce `BindingService.getProducerBinding(String bindingName)`
and `getConsumerBindings(String bindingName)` to cover a use-case
when we need to check for predefined bindings.
This API might be useful in other use-case when we need to get
access to managed bindings

Related to: https://github.com/spring-cloud/stream-applications/issues/330