KafkaBinderConfigurationProperties metadata
When generating configuration metadata (spring-configuration-metadata.json),
KafkaBinderConfigurationProperties class is missing all the information except
for the constructor bound KafkaProperties. We need to add the @Autowired on the
constructor so that Boot will not do a constructor binding and thus include all
the JavaBeans style properties in the metadata. See this Boot issue for more details.

https://github.com/spring-projects/spring-boot/issues/34031

In addition, this @Autowired is necessary for the properties to be correctly
bound when running in native mode.

Resolves https://github.com/spring-cloud/spring-cloud-stream/issues/2640
Resolves https://github.com/spring-cloud/spring-cloud-stream/issues/2644