Revert the #2989, it may cause unexpected results.
In spring-core `GenericConversionService` contains dozens converters and if nothing is matched, the `NoOpConverter` will be used. The `SpelConverter` of Object -> Expression is too general, and cause for some type conversions instead of picking `NoOpConverter` the `SpelConverter` be selected.

Using String.valueOf(object) is not safe and for classes that do not override `toString` or their `toString` is not Spel compatible cause to throw Exception.
e.g., LiteralExpression -> Expression used NoOpConverter but after the #2989 change it will pick SpelConverter and causes problem.
It is better to save the previous application.yml behaviour for expressions and change `spring-core` if this functionality is needed.

Signed-off-by: sinaaskarnejad <sina.askarnejad@gmail.com>